requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
TTS>=0.22.0
pillow==10.1.0
ffmpeg-python==0.2.0
apscheduler==3.10.4
pydantic==2.5.0
python-dotenv==1.0.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
numpy==1.24.3
librosa==0.10.0
scipy==1.11.4
wavefile==0.0.14
requests==2.31.0
feedparser==6.0.10
groq==0.4.2